refactor(nextstep): Direction 객체 이동 및 getter 추가
